requirements.txt
streamlit
psycopg2-binary
SQLAlchemy
langchain
langchain-community
openai
python-dotenv
pandas
